ref(tiller): refactor tests into logical files
TestListReleasesByStatus -> tiller/release_list_test.go
TestListReleasesFilter -> tiller/release_list_test.go
TestListReleasesSort -> tiller/release_list_test.go
TestListReleases -> tiller/release_list_test.go
TestUpdateRelease_ResetReuseValues -> tiller/release_update_test.go
TestUpdateRelease_ResetValues -> tiller/release_update_test.go
TestUpdateRelease_ReuseValues -> tiller/release_update_test.go
TestUpdateReleaseNoChanges -> tiller/release_update_test.go
TestUpdateReleaseFailure -> tiller/release_update_test.go
TestUpdateReleaseNoHooks -> tiller/release_update_test.go
TestUpdateRelease -> tiller/release_update_test.go
TestRollbackReleaesFailure -> tiller/release_rollback_test.go
